Finish the Observable for the Gameboard
This commit fixes the Null Pointer Exception thrown when the method
that initializes the ObservableArrayMap with values based on the initial
configuration of the Gameboard. There was a error in the logic of the
Gameboard's constructor where instead of instantiating 'empty' Squares
in the array, some of the Squares with no Pieces weren't getting
instantiated. This was causing a NPE to be thrown when the loop in
the CheckersViewModel class was checking if the Square was empty.

After fixing this I finished the connection between the Observables and
the View by setting the background of the buttons where pieces will go
using the value for the key associated with this button in the
ObservableArrayMap. I used nested ternary operators to accomplish this.
If the value for that button's key in the Observable is 0 (empty) the
background will be white, if the value for that button's key is 1
(black) the background will be white with a black piece shown, and if
the value for that button's key is 2 (red) the background will be white
with a red piece shown.

Android Data Binding Documentation:
https://developer.android.com/topic/libraries/data-binding/index.html

Example of Android MVVM Design Paradigm using Data Binding:
https://academy.realm.io/posts/eric-maxwell-mvc-mvp-and-mvvm-on-android/