Allow highlighting of a piece's available moves
I implemented the ability for a square on the checkerboard to be tinted
a certain color to indicate that it is a valid move for the selected
piece. The button's backgroundTint attribute will change based on the
boolean value of key associated with that particular button, which is
stored in an ObservaleArrayMap in the ViewModel.

It was necessary to also set the backgroundTintMode attribute so that
the normal background is visible through argb transparancy instead of
being totally obscured.

Android backgroundTint attribute Documentation:
https://developer.android.com/reference/android/view/View.html#attr_android:backgroundTint