Add a position instance variable to Piece
I had previously moved the position instance variable to the Square
class when I did the MVVM refactor and changed the implementation of the
GameBoard.

Storing a Piece's position in the object makes it much easier for the
search algorithm the Bot is using to work. The Bot is able to iterate
over the sets of Pieces in the GameBoard instead of having to traverse
the 2D Square array searching for pieces while remembering the Square's
position.