Add a method to clone a GameBoard
The GameState method next() that the bot uses, iterates over every piece
for a player, executing every available move for that piece on the
GameBoard. To ensure on iteration does not interfere with the GameBoard
used in another iteration, the GameState method clones the board first
and then operates on the cloned board.