Implement next() for the Bot's use
I added a instance variable for the state so that I could impose a limit
on the amount of states the Bot could explore before making its
decision.

I then implemented the first next() method which returns a collection
of every possible state that can result from the current state. The
Bot depends on this method. I partially implemented a way to determine
if the Game is over in a particular state (it needs more work).