Add a method to determine the winner
Removed the deprecated play() method. Added a method used to determine
the winner of the game. This method along with the one in GameState to
see if the state of the Game is over need a bit more testing and work.

I also had to reset the Child State counter of the GameState instance
or the Bot wouldn't be able to make another move. The counter is only
supposed to limit the amount of states to be expanded in one turn, not
for the entire game.