Add the Bot to the ViewModel's logic
Currently, in order for the Bot to be used, a new Bot instance must be
passed to the Game constructor in onNewGameSelected().

I added logic to the Click handling method to ignore clicks that are
not selections of available moves, and to allow for the selection of a
different piece if the piece is clicked twice.

I also added in the logic that determine if the opponent is a Bot or a
Human, and prompt the Bot to take its turn if there is a Bot.

I renamed the method initializeGridAndAvailableMovesObservables() to
syncGridAndAvailableMovesObservables() to better describe how its used.