Highlight selected piece during move
When a piece with no available moves was selected, there was no feedback
on in the view. Because of this it was hard for the user to tell what to
do if they selected a piece with no available moves and then tried to
select another piece before de-selecting their piece, the app would seem
unresponsive.

This commit highlights the selected piece which adds more meaingful
feedback for the user while clicking.

Closes #12