Make Restart Game logic Null-safe
Because the Restart game, assumed there was a game already in progress
when you clicked it, it did not perform check to see if the Model
existed before trying to clear it. This resulted in a NPE.

I moved all of the Restart game logic into if statement that checks
if the model is null.

I also added calls to update the Observables for captured pieces in both
the Restart game method and the New game method in the ViewModel.

Fixes #7 and fixes #15