Add Logic to Determine the End of the Game
I added logic to the Model to determine when the game is over. The game
is over when all of a player's pieces are captured or they have no more
available moves.

I added a method in the ViewModel that produces a Game Over popup when
the GameState is over and a winner has been determined. The popup
allows the user to restart the game or clear the board.

Closes #21