Clear Selected Piece from Human during Restart
Sometimes the app would crash and other times the wrong piece would
move after selecting Restart Game in while a move was in progress. This
occured because although I was clearing the state of the Game instance,
I had forgotten that the selectedPiece instance variable is stored in
the Human instance and not the Game. It was not getting cleared.

To fix, in the onRestartGameSelected method I just check if either
player is a Human and if so, I clear their selectedPiece.

Fixes #8