Add better logic for the Winner calculation
This commits adds logic to the GameState to flag the state as 'over' if
the current player has no more available moves. In the Game class
this method is used to determine the winner. If the GameState is 'over',
then one of the players cannot move, so the player without any more
moves is the loser and the other is determined to be the winner.