Fix app crashes from clicking on overlay screen
The overlay screen appears before a game Model is instantiated. The
On Click callback method referenced a method of this model and this
was causing a null pointer exception.

I wrapped the entire On Click logic in an if statement checking to make
sure the model was not still null.

I also changed the number of expanded states back to 10,000 from
50,000 because the game would freeze for over a minute sometimes when
the bot was making its decision.

Fixes #25